requirements.txt
google-cloud
google-cloud-bigquery
google-cloud-bigquery-storage
protobuf
pandas